Refactor early-loading UI code so that less code is loaded by stats server
