+++ .idea/runConfigurations/Stats_Server.xml
@@ -2,7 +2,7 @@
   <configuration default="false" name="Stats Server" type="TypeScriptProgramRunner" factoryName="TypeScript">
     <module name="gearplan" />
     <envs>
-      <env name="SHORTLINK_SERVER" value="http://localhost:8081/shortlink/" />
+      <env name="SHORTLINK_SERVER_F" value="http://localhost:8081/shortlink/" />
     </envs>
     <option name="interpreterRef" value="project" />
     <option name="enabledTsNodeEsmLoader" value="false" />
