+++ packages/frontend/src/scripts/components/saved_sheet_picker.ts
@@ -1,8 +1,8 @@
 import {CustomCell, CustomColumn, CustomRow, CustomTable, SpecialRow, TitleRow} from "../tables";
 import {SheetExport} from "@xivgear/xivmath/geartypes";
 import {faIcon, makeActionButton} from "./util";
-import {getHashForSaveKey, openSheetByKey, showNewSheetForm} from "../main";
 import {deleteSheetByKey} from "../persistence/saved_sheets";
+import {getHashForSaveKey, openSheetByKey, showNewSheetForm} from "../base_ui";
 
 export class SheetPickerTable extends CustomTable<SheetExport, SheetExport> {
     constructor() {
